disk: NewCustomPartitionTable() function
New function that creates a partition table from scratch based on the
new disk customization.

First, it creates any necessary partitions for the selected boot mode
(BIOS boot, ESP, or both).
Then it determines if a /boot partition is needed based on the
partitions and mountpoints defined in the disk customizations.
Then, it iterates through partitions and creates each one in the order
its defined.
Finally, it creates a root filesystem if one wasn't already defined,
resizes any necessary volumes, relayouts the table (setting all
partition starts and offsets), and generates UUIDs all entities.

Because of the new way the boot partition is handled, the
EnsureBootPartition() function was changed to always add a boot
partition and was renamed to AddBootPartition() to match the new
behaviour.  The tests have been updated accordingly.