osbuildmonitor: include a "Trace" string in Status
This commit adds a `Status.Trace` string in the osbuildmonitor.

Trace contains a single log line, usually very low-level or
stage output bug useful for e.g. bug reporting. Should in
general not be displayed to the user but the concatenation
of all "trace" lines should give the same information as
running osbuild on a terminal.