osbuild: add support for jsonseq monitor/status message
This commit adds support for scanning/parsing the jsonseq based
monitoring that osbuild provides. This is useful for machine
readable status reporting and progress information for e.g.
`bootc-image-builder`.

The API consists of a new `osbuild.Status` struct that contains
parsed details about the osbuild status/progress and a scanner
that can scan a stream of osbuild jsonseq based progress and
emit `osbuild.Status` structs.

Note that the status is broadly of two categories:
1. information amining at a user interface (like
   `status.Message` or `status.Progress`)
2. low-level details for logging/error reporting (like
   `status.Trace` and `status.Timestamp`

Note that error handline a bit weak right now. The consumer
of the API will not get an error via the monitor currently
but only know that something went wrong from osbuild exiting
and then the consumer needs to reconstruct the error log from
the trace messages. This is not a show-stopper but we could
improve this by merging https://github.com/osbuild/osbuild/pull/1831
and adding `Result` to the `statusJSON` and `Status` (but this
needs more thinking and probably best done in followups).

See https://github.com/osbuild/bootc-image-builder/compare/main...mvo5:progress3?expand=1 how this will be used.