blueprint: decouple FilesystemCustomization from marshaling
This commit decouples the marshaling representation of the
`FilesystemCustomization` from the actual struct. This means that
we can reuse custom unmarshaling via datasizes.Size and also
have our external represenation "pure".

If we continue using this pattern is also means that we can rename
fields in the marshaling and provide compatbility easily.

Thanks to Thozza, c.f.
https://github.com/osbuild/images/pull/1049#issuecomment-2493261642
and
https://github.com/osbuild/images/pull/983