blueprint: make `minsize` required for `disk.Customizations`
This commit enforces the use of "minsize" for disk customizations.
Without this the `disk` code will create zero sized partitions when
`minsize` is omited which then later yields a strange error from
osbuild:
```
Building manifest-qcow2-raw-vmdk-vhd-gce.json
<stdin> has errors:

.pipelines[1].stages[2].options.volumes[0].size:
  '0B' does not match '[1-9][0-9]*[bBsSkKmMgGtTpPeE]?'

2024/11/25 09:35:42 error: cannot run osbuild: running osbuild failed: exit status 2
```

We could also attempt to fix this by making guesses about minsize
(which is what the existing filesystem customizations do). But that
seems less appropriate here as this is about finer control and an
arbitrary size of e.g. `1 GiB` seems unexpected. Another alternative
would be to infere from the parent container size and split equally.
But we can always do this later and relax the rules (and even add
things like `minsize: 50%` which would imply the parent container).