RHEL-10/AMI: sync function names with RHEL-9
RHEL-9 AMI has been reworked as part of the RHSM auto-reg v2, which
resulted in all RHSM and RHUI specific parts to be removed from the
image definitions. This allowed to simplify them and some functions
were renamed as part of it.

Carry over the function names used in RHEL-9 to RHEL-10, to make things
more consistent across these distributions.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>