LoadAllRepositoriesFromFS: log only filenames of loaded repositories
After the rework done by PR#1038 [1], the `configFile` changed from the
file path string to `fs.File` instance. As a result, logging it makes
the output verbose, as it logs the whole content of the loaded file
(repo URLs, GPG keys, etc.). Revert to the old behavior by logging
`fileEntry.Name()` for the loaded repository.

[1] https://github.com/osbuild/images/pull/1038

Signed-off-by: Tomáš Hozza <thozza@redhat.com>