disk: move TestPartitionTables to internal/testdisk
Move the TestPartitionTables from the disk tests to the
internal/testdisk package.

This makes the test partition tables reusable across the project, but
also makes it impossible to use them in internal disk tests (import
cycle).  The PartitionTableFeatures test is therefore moved to the
disk_test package instead and the function is exported for testing in
the pkg/disk/export_test.go file.