osbuild: rewrite GenMkfsStages() to GenFsStages() and handle swap
Replace the GenMkfsStages() function with a more general GenFsStages().

This new function iterates through all entities, not just mountables,
and handles, in addition to filesystem creation (org.osbuild.mkfs.*),
btrfs subvolume creation (org.osbuild.btrfs.subvol) and swap
(org.osbuild.mkswap).

This removes the need for calling the separate GenBtrfsSubVolStage()
when generating image pipelines.

This change has a minor effect on the org.osbuild.btrfs.subvol stage
options in manifests: it changes the name of the device (which is not a
functional change) and locks the loopback device (which is not necessary
but shouldn't cause issues).

Tests have been updated accordingly.