disk: new interface: FSTabEntity
A new interface FSTabEntity which describes entities that can appear in
the /etc/fstab file.  This new interface replaces most of the Mountable
interface, in which it is now embedded.  The new interface will be
useful to separate entries that cannot be mounted but do appear in the
fstab file (namely, swap areas and swap files).

All current Mountables (Filesystem and BtrfsSubvolume) implement
GetFSFile() now as an alias to GetMountpoint().