disk: add ForEachFSTabEntity() function
This function is the same idea as ForEachMountable() but for the new
FSTabEntity interface.  The new function is used in the
NewFSTabStageOptions() now which is more appropriate.  The distinction
will be useful with the addition of swap areas that should appear in
/etc/fstab (and therefore, the org.osbuild.fstab stage options) but
should not be considered for the mkfs stage generation.