disk: cover all entities in PartitionTable.features()
Add a case for every entity type in the PartitionTable.features()
callback.  More importantly, panic if any entity type is not covered by
a case.  This ensures that when a new entity type is added, we need to
consider whether it needs to be represented as a feature, added to an
existing one, or ignored.

Co-authored-by: Michael Vogt <michael.vogt@gmail.com>