distro: add tests for ESP partition availability
TestESP checks whether all images marked as UEFI or hybrid bootable have
an ESP partition in their base partition table. The opposite is
checked as well: legacy-only images shall not have an ESP partition.

The code is deduplicated as much as possible. It meant adding a proxy
function argument to the shared TestESP function because partition
tables are not a public interface, but it's imho good enough.

Maybe we can go one step further once the GetPartitionTable interface
is unified across rhel and fedora, and just make it public. Maybe.