disk: align logical volumes up in EnsureSize()
When creating an LV they need to be aligned up boundaries
or lvcreate will fail. When a lv is create via CreateLogicalVolume()
this hapens because the size is calculcated with `v.AlignUp()` but
when a resize is required via `lv.EnsureSize()` this is not taken
into account.

This fixes the failure in bib PR#750