osbuild: validate sfdisk stage options
DOS partition tables only support up to 4 partitions.  Don't create the
sfdisk stage options if this rule is violated.  Stage validation
failures panic, so we panic on the stage creation.  This might change in
the future (to return errors) but for now we use panics and put the
responsibility on the callers to validate their inputs if they want to
have better error control.