disk: error on PTs with type dos and len(partitions) > 4
Return an error when a partition table is created with type dos and more
than 4 partitions.

This check can only happen after the partition table is created,
otherwise we would have to predict all the partitions that get created
automatically, which would duplicate a lot of the logic in
NewCustomPartitionTable().

Update tests:
- Convert the plain+swap (valid) test to GPT since now it's invalid as
  DOS.
- Add a fail test for a DOS partition table with too many partitions.