disk: fix partition type IDs for dos
The correct partition type ID for ESP is 'ef' (not 'ef00') [1].
The partition type ID for LVM is '8e' [1].

BIOS boot partitions aren't meant for (or required) on dos partition
tables [2].  We will eventually make this a partitionless gap in the PT,
but for now let's simply use the code for 'empty'.

Use the getPartitionTypeIDfor() function for all partition type
assignments when creating partitions.

Add tests that cover all basic partition table configurations (plain,
lvm, btrfs) using dos and gpt.

[1] https://tldp.org/HOWTO/Partition-Mass-Storage-Definitions-Naming-HOWTO/x190.html
[2] https://www.gnu.org/software/grub/manual/grub/html_node/BIOS-installation.html#BIOS-installation