tests: Update TestImageTypePipelineNames to be less strict
This test assumed the image's static list of pipeline names must match
the manifest, but it really only needs to check the ones with rpm or
ostree metadata.

To support the ability to have images with different pipelines (eg.
squashfs only vs. ext4+squashfs rootfs iso creation) this changes the
test to make sure the pipelines in the static list are in the manifest,
and in the order specified. But the two lists do not need to match
exactly.