distro: Remove rootfs-image from fedora pipeline lists
This pipeline is no longer used on Fedora 41 and later. They use a plain
squashfs install.img instead of wrapping an ext4 filesystem with
squashfs.

But it may be present in older versions, the test in distro_test.go has
been adjusted to ensure that all pipelines named are present and in the
correct order without requiring that this is an exact list of pipeline
names.