pkg: make `seed` in `ImageType.Manifest()` random by default
This commit changes the API of `ImageType.Manifest()` to take
a pointer to the seed value. If no seed value is given it will
automatically generate a random one.

This avoid the pitfall that the consumer needs to know about
properly seeding the manifests. The default use-case is to
generate non-fixed seed manifests so this commit it easier.