data: copy all repositories from `osbuild-composer` to images
This commit moves the repositories from `osbuild-composer` to
images. The rational is that we need a central place to host
the repositories so that they can be shared between osbuild-composer
and image-builder-cli.

The other reason is that conceptually images is (arguably) be a
better place because it knows how to build osbuild manifests and
it can only build those for distro releases it knows about. The
filenames are already part of the API of images (as it derives
the distro name from the filename) and it is impossible to add
new distros (except maybe point releases) without also updateing
images to deal with the details of the new distro.