disk: use custom partition type (part_type) when making partition table
Read the custom part type if it's set, otherwise infer it from the
mountpoint and partition table type.

At this stage, in NewCustomPartitionTable(), we know the partition table
type, so we can verify the partition type ID again with more strictness.

To avoid duplicating the validator, we use the partition method again.
It would have been nicer to have this kind of validation in the disk
package instead of the blueprint, but that would create an import cycle.