disk: enable customizing ESP partitions
This commits allow people to customize their ESP partition. Previously,
we always added an ESP partition ourselves, regardless of the given
customizations, so the customization errored out if there was an ESP
because of a conflict. This commit alters the behaviour, so ESP is
only added automatically if it's not in the customizations.