Test/imgtestlib: make get_osbuild_commit() consider only gitlab runner
Previously, the osbuild commit stored in image build `info.json` was
based on the host distro. However, we always use the same distro for all
image build jobs, which is defined in the Schutzfile.

Modify the `get_osbuild_commit()` function to not take the distro as an
argument, but instead read the `gitlab-ci-runner` from the Schutzfile
and determine the osbuild commit based on its distro.

This will make it much easier to determine the actual osbuild commit
that was (supposed to be) used for the image builds on the checked out
branch. Consistency of data in `image.info` and the reality will be
ensured.

Make `get_osbuild_commit()` raise a KeyError in case there's no
osbuild commit defined in Shutzfile for the gitlab-ci-runner distro, so
that we do not accidentally miss it and use unknown osbuild version from
the distro repos.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>