update-schutzfile-osbuild: update commit in all distos
Modify the script to not only update osbuild dependency commit for all
distros defined in the Schutzfile, but also define it for any distros
that didn't have it.

This will ensure consistency of using a well known osbuild version when
testing image builds in CI and storing it in the S3 CI cache with the
osbuild ref being part of the object prefix.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>