GH actions/test: don't setup repos outside of the container env
I noticed that the action that is running unit tests on Fedora was
setting up YUM repository for osbuild on the Ubuntu runner and then
again inside the Fedora container in which it runs unit tests.

Setting the osbuild YUM repository on the Ubuntu runner is useless,
therefore delete the task.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>