Test/imgtestlib: add runner distro to the build cache S3 prefix
Add the runner distro version to the image build cache S3 prefix. This
is a preparation for not using osbuild NEVRA, but osbuild commit ref
and used CI runner distro. The goal is to make it easy to determine
the S3 prefix corresponding to checked out branch without the need to
setup YUM repo and install RPMs on the CI runner.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>