Test/imgtestlib: replace osbuild ver by osbuild ref in the S3 prefix
Replace the osbuild nevra in the image build cache S3 prefix with
osbuild ref and prefix it with "osbuild-ref-" to make the meaning of
the SHA clear. The goal is to make it easy to determine the S3 prefix
corresponding to checked out branch without the need to setup YUM repo
and install RPMs on the CI runner.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>