Test/imgtestlib: introduce gen_build_info_dir_path_prefix()
Let's unify the generation of build info dir path prefix into one
function, instead of generating it in various ways in multiple places.
This commit introduces a new function to generate the relative dir path
prefix, depending on provided arguments.

The existing code that generates the prefix on its own will be replaced
in the following commits.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>