imgtestlib/gen_build_info_*: don't require osbuild_ref and runner_distro
Modify `gen_build_info_dir_path_prefix()` and
`gen_build_info_s3_dir_path()` functions to not require osbuild_ref and
runner_distro to be passed explicitly as arguments. If not provided, the
runner_distro defaults to the host distro and the osbuild_ref defaults
to the pinned version for the runner_distro. This simplifies all the
places which call these functions.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>