manifest: change signature of `manifest.Serialize()`
This commit changes the signature of `SerializeFull()` to take a new
`Inputs` struct and a new `Options` struct. The `inputs` make it
easier to extend the function as for new inputs the signature
no longer needs to change. And `options` allows us to pass extra
paramters like what rpm downloader to use.

Note that this keeps a compatibility Serialize() for now, mostly
to avoid churn in this PR but we definitely want to clean this
up and fix the callers.