osbuild: add rpmDownloader to `GenSources()` argument
This commit enables to select what rpm downloader backend to use.
The signature of GenSources() is just expanded to take an additonal
parameter. It would be nice to re-use the `manifest.Inputs` struct
but it is (subtly) different from what GenSource needs :/ I was
also tempated to add a "Options" struct to GenSources instead of
the enum but YAGNI - if we ever need to add another one this will
become a options struct.