many: add "ref" in file customizations to support host resources
We have a feature request that asks to support including local
files in blueprints. This also came up recently in a dicussion
about the blueprints v2 format (that is where the name "ref" comes
from). This is tricky of course for the service and would probably
not be supported there initially (maybe later via a extra upload
option). But for `image-builder-cli` it does make sense and is
a useful feature that is support by other image-building tools
too.

This initial commit show how we can use blueprints and a new "ref"
field to give access to host files [0]. It works internally by
generating an osbuild manifest with `org.osbuild.files` inputs
that get populated via a `curl` stage [1] that pulls the files
from the host into the osbuild store. Because they are addressed
by hash (just like inline files) the manifest is still reproducible
and if anything changes in the file(s) from the host the manifest
will fail to build. Note that all of this is an implementaion detail
and we could change "osbuild" to do something else, the blueprint
API would not change.

Note also that this means that an attacker could potentially use
this in combination with custom rpm packages and maintainer scripts
to gain information about files outside the buildroot. As our isolation
is not a security feature this shouldn't be an issue (and arguably
running maintainer scripts in the rpm stage as root is already
opening a lot of possibilities to break out of the buildroot).
This is only an issue when the inputs are untrusted, e.g. in the
service in the future.

Blueprints generated like this are no longer portable as they
require the local files and when those change the manifest breaks.

They will become portable when we implement "offline" bundles,
i.e. a way to generate a manifest and their dependencies in
a single file that can then be rebuild by image-builder/osbuild.
This implementation works nicely with that idea because in the
most simple case a "offline-bundle" is just a manifest+store
directory. Because everything is looked up in the store first
by hash this will "just work". Offline bundles will also
fix the other issue with manifests that if sources like packages
go away (because they are outdated) the manifest also stops
building. Note that we can even have "half-offline" bundles
where we would put only the "file" refs into the .store we
generate but keep the package urls out.

The downside of using the store of course is that it means
duplication of data and may seem wasteful for big files.
However without that it is very hard to keep the gurantee
that a manifest is reproducible because if we access the
file directly we have a race-condition that the hash may
change anytime. Any trick like bind-mounting or hardlinking
will have the same issue. So unless we drop the reproducibility
for local file refs (which if of course an option) I see
no other way than to have a copy in the store.

[0] Initial commit because only regular files are reported
for now. The mechanism is generic enough so that we could
support dirs (via simple recursion) and http{,s} URLs as
well (actually anything that curl supports but we should
currate it carefully).
[1] IIRC coreos uses this trick as well in their manifests

Jira: RHEL-21334