manifest: change the serializeStart() to take a new `Inputs` arg
The signature of serializeStart() is a bit unwieldy and when
adding librepo based sources this prompted for this proposed
cleanup. The idea is to just pass a single "Inputs" struct
to the serialization so that if we add field we don't need
to change the signature all over the place. It also makes
the calls easier to read (IMHO).