manifest: add `SerializeOptions` to `manifest.Serialize()`
This commit makes `Serialize()` more extensible by switching the
rpmDownloader option to a more general `SerializeOptions` struct
that can be nil if no options are required. It contains the
`RpmDownloader` as the only option right now. We could YAGNI it
and do it once we have the second option but arguably it is cleaner
already doing it now.