manifestgen: add sbom support via SBOMWriter() callback
This commit adds an `SBOMWriter()` callback to the
`manifestgen.Options` struct. This allows to export
the SBOMs when the manifest is generated.

Note that usually at least two files are generated,
one for the "image" payload and one for the "buildroot",
but it can be an arbitrary number.

They typcially look something like:
```
centos-9-qcow2-x86_64.buildroot-build.spdx.json
centos-9-qcow2-x86_64.image-os.spdx.json
```
The callback will not write the files directly but just
provide the suggested filename and the data so that the
caller is flexible and can also put it in e.g. a database.