image: allow setting the locale for Installer images
Support setting the locale on the top-level installer image and copy it
down into the pipeline.
Adds a new 'Locale' property to most installer image kinds: Live,
OSTree, and Container.

The image (tar) installer already has OSCustomizations, which define the
language/locale for the payload to be built.  The ISO reuses that same
locale for consistency between installer and payload.