manifest: create groups before users in ostree deployments
In regular, package-based builds and in base ostree commits, we always
create groups before creating users, i.e., we always run
org.osbuild.groups before org.osbuild.users.  This is the correct way to
do it because a user in a blueprint can have a gid that is defined in
the group section of the blueprint.  In ostree deployments (e.g.
edge-qcow2/iot-qcow2 images) this scenario would make the build fail,
because we were creating users before groups.

Flip the order to fix the bug.

Thanks to Kenric Young for discovering this bug and helping pinpoint it.