github: add CODEOWNERS
Add a CODEOWNERS file that makes PRs get automatically assigned to
people base on the package path or file globs.

For now, I added the new osbuild-reviewers team as a global default.
Random people from that team will be assigned when a PR is opened and
doesn't match a more specific rule.

I expect more rules to be added in the future by people wanting to take
specific ownership of certain packages or paths.

See https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#example-of-a-codeowners-file