dnfjson: add DepsolveAll() convenience function
A convenience function that runs Depsolve() on each package set slice
and returns a map of results with the same keys.

This is a follow-up to 5abe1d55f12fa61ecfc041d4ac30c93a66e2f2f0 which
introduced the DefaultDepsolver() function in manifestgen.  It
essentially moves the core of that function into the dnfjson package.