distro: use `rand.Int63` intead of `time.UnixNano()`
This is a slightly silly commit. Instead of using `time.UnixNano()`
its much more sensible to use `rand.Int63` which is automatically
seeded since go1.20 [0] and also gives us a wider range of possible
seeds than UnixNano (which will always be of a preditable length).

None of this really matters in practise as we just need something
"good enough" but using rand.Int63() is the correct thing to do.

[0] https://tip.golang.org/doc/go1.20#mathrandpkgmathrand