imagefilter: make `short` output format as unstable
This commit changes the "short" output format so that it is
marked as "unstable". There are some open questions about this
format and how it relates to e.g.
https://github.com/osbuild/osbuild-composer/pull/4336
which adds very similar API is a slightly different format.

So to prevent users from relying on the API before that is resolved
we need to at least make them aware that this may still change.