container: new resolver: blockingResolver
The blockingResolver implements the Resolver interface but resolves
containers synchronously, blocking execution on the resolve process
inside the Add() method.

This will become the only implementation in the near future.  For now we
will support both to make the transition easier for osbuild-composer.

This change is made so that the container resolver behaves in the same
way as the other content resolvers (packages and ostree commits) for
consistency.