imagefilter: use `distrosort` to sort the distro names
This commit uses the `distrosort` package to sort the distro names
by their name/version. This means that the output of centos is sorted
by low to high version like this:
```console
$ ./image-builder list-images --output=short --filter distro:centos*
centos-9:
  ami: [ aarch64, x86_64 ]
  ...
centos-10:
  ami: [ aarch64, x86_64 ]
  ...
```

with the regular "string" search this was:
```console
$ ./image-builder list-images --output=short --filter distro:centos*
centos-10:
  ami: [ aarch64, x86_64 ]
  ...
centos-9:
  ami: [ aarch64, x86_64 ]
  ...
```
(this is also quite visible with rhel).