container: add Resolve() and ResolveAll() to resolver
Extend the Resolver interface to implement Resolve() and ResolveAll().
This is consistent with the API of the other resolvers.

The blockingResolver implements Resolve() as its main resolver function.
The Add() function now wraps Resolve() and collects the resolved specs
in the results slice.  This is the inverse of the asyncResolver where
Resolve() wraps Add().
ResolveAll() calls Resolve() iteratively on all container source specs.