reporegistry: expand `New()`, unexport `loadAllRepositories`
This commit changes `reporegistry.New()` to take a new parameter
`repoConfigFS []fs.FS`. With that we can unexport the public
`LoadAllRepositories{,FromFS}` as the only user for those is
`composer` and that can simply use `.New()`.

Note that we could also change `New()` to just take a single
`repoConfig []fs.FS` but the downside of this is that the
error reporting becomes more awkward, i.e. to report what
paths where configured would (probably) require to build
them manually via `fs.FS.Open(".").Stat().Name()` (but
I have not tested this).