osbuild: define the minimum osbuild version and enforce it
Define a new variable storing the minimum osbuild version supported
by the module and an exported function to get it. The actual version is
stored in a text file under `data/dependencies/` directory and embedded
at compile time. Enforce the minimum version when executing osbuild.

This has the potential to prevent surprises when using older osbuild
with `images` implementation that relies on changes in a newer version
of osbuild.

This will allow dependent projects to extract the minimum version and
check if their packaging requirements are up to date with the version of
`images` that they depend on.

This will enable adding a composite GH action that can be used by other
projects depending on `osbuild/images` to check that their SPEC file
depends at least on the specified osbuild version.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>