osbuild: define the minimum osbuild version and enforce it
Define a new constant in the `osbuild` package for the minimum osbuild
version supported by the package. Enforce the minimum version when
executing osbuild.

This has the potential to prevent surprises when using older osbuild
with `images` implementation that relies on changes in a newer version
of osbuild.

This will allow dependent projects to extract the minimum version and
check if their packaging requirements are up to date with the version of
`images` that they depend on.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>