osbuild: generate systemd mount stages from partition table
GenSystemdMountStages() generates systemd .mount and .swap stages based
on a partition table.  This is an alternative to generating an fstab
stage for the mountpoints.

The generator creates one systemd.unit.create stage for each mountpoint
(and swap) and an extra stage that enables all the new units after they
are written.

vfat volume IDs are discovered by matching against a new regular
expression defined in the `mkfs_fat_stage.go` file.  This is needed
because non-UUID identifiers are not lowercased in /dev/disk/by-uuid.