imagefilter: use consts to avoid "loose" strings
This commit moves the "loose" strings for the filtering into
consts. Thanks to Simon for the suggestion.

Fwiw, we could also move this [0] into a
```go
var supportedFilters map[string]prefixMatchFunc{
    "distro": matcherFuncDistro,
    ...
}
```
which would mean the supportedFilters function and the actual
supported things can never go out of sync. The downside is that
we would still need to maintain the order of the supported filters
manually as some are just way more useful than others so naively
sorting goes not give the best user experience.

[0] See
https://github.com/osbuild/images/commit/5e50e9346bcd8fafc9783e844cce08f74ba80a2f
for a full version.