distro/packagesets: define package set loader function
The Load() function reads the package sets for a given distribution and
selects the specific set for the image type based on the image type
name.

Conditions are processed if any are defined.

The 'replacements' argument defines a set of string replacements, which
currently only apply to version strings.  This will be used to replace
version names like VERSION_RAWHIDE with the value defined by the
distribution.