cloud: add `AwsUploader.Result()` with strutured info
This commit adds a new `AwsUploader.Result()` call that allows to
get structured information about the upload result. This is needed
for e.g. osbuild-composer.

It is not part of the `UploadAndRegister()` call because the exact
information about the result is cloud specific. Having a common
uploader interface is nice as it simplifies/unifies the common code
paths. So this tries to be find the right balance. Just like each
uploader will have a different `NewUploader()` with different
arguments the `Result()` will also be slightly different for them
all.