distro: move TestManifestRepositoryCustomization into distro
This commit moves the TestManifestRepositoryCustomization into
the distro package to avoid the previous duplication.

The test location is not ideal, but either we duplicate this
test in rhel/imagetype_test.go and fedora/imagetype_test.go
or we have one version here that has knowledge about the
details of how images types are implemented in
{fedroa,rhel}/imagetype.go (i.e. how workloads and payload
repos get added during the imagetype.Manifest() call).