platform: add Bootloader option to the x86_64 platforms
Bootloader is an enum type with three values:
1. None
2. GRUB2
3. ZIPL
4. UKI

While technically not a bootloader, UKI is defined as one of the options
because it replaces the bootloader and it is used to determine
bootloader package selection.

For now, None defaults to GRUB2.  We don't want to have to set the
bootloader for all x86_64 platform instances yet.