manifest: always apply selinux labeling in bootc
This commit runs selinux labeling unconfidionally. Traditionally
we wanted to touch the filesystem that bootc created as little
as possible. However as Achilleas rightfully pointed out we
always write the /etc/fstab so we always need to label.

If we go back to not write /etc/fstab and use kargs to find the
rootfs (see issue#756) we probably should to revert this again
and make labeling conditional again.