lib: add new `experimentalflags` package
This commit adds a new `experimentalflags` package that can be
used to enable experimental features in the image-builder ecosystem.

The initial implementation provides a `Bool()` function that allows to
do e.g.:
```
IMAGE_BUILDER_EXPERIMENTAL=skip-arch-checks
IMAGE_BUILDER_EXPERIMENTAL=skip-arch-checks=true
```
to disable architecture mismatch checks. This will be used as part
of the `image-builder-cli` but the mechanism is generally useful.

And also a `String()` implementation that just passes the verbaim
string on.