manifest: select bootloader stages based on platform.Bootloader
Use the platform.Bootloader property to select bootloader stages.
Importantly, when the platform isn't configured to use grub2 as the
bootloader, don't add any grub-related stages.

For s390x, we use the bootloader property to select the zipl stage,
instead of basing the decision on the architecture itself.  In reality
this has no effect on our images or manifests, since all our (bootable)
s390x images have the bootloader set to zipl.