cmd/gen-manifests: sort package sets in mock depsolver
Since the packages were extracted into yaml files in #1205, the package
list order has become unstable, in particular when packages are added
conditionally.  I'm not entirely sure where the instability comes from,
but generating manifests with the mock depsolver creates lists with
different order for the OS pipeline for the image types that
conditionally add "dnsmasq".

Sorting the package lists ensures that manifests don't change from this
instability.  It also makes the manifests reproducible when the package
lists are reordered internally, which we would prefer to have, since the
package order pre-depsolve shouldn't affect the final manifest.