distro: define ImageType.KernelOptions as a slice
Make KernelOptions a string slice where each element is a separate
option.  This makes it easier to manipulate each option individually,
which we will need later to modify default options when using mount
units.

This commit has a non-functional effect on manifest generation:
Edge and IoT images that use org.osbuild.ostree.deploy stage now define
each option as a separate element in the `kernel_opts` slice.
Previously, it would be a mix of space-separated multi-option strings
and individual ones.